Moving the sample application to sample-application/.
