+++ .npmignore
@@ -1,4 +1,4 @@
-Apps/
+sample-application/
 swagger-ui-*/
 .gitmodules
 *-project
