+++ sample-application/main.js
@@ -19,9 +19,9 @@
 var express = require("express")
  , url = require("url")
  , cors = require("cors")
- , swagger = require("../../lib/swagger.js");
+ , swagger = require("../lib/swagger.js");
 
-var petResources = require("./petResources.js");
+var petResources = require("./resources.js");
 
 var app = express();
 
@@ -104,7 +104,7 @@ swagger.configureSwaggerPaths("", "api-docs", "")
 swagger.configure("http://localhost:8002", "1.0.0");
 
 // Serve up swagger ui at /docs via static route
-var docs_handler = express.static(__dirname + '/../../swagger-ui/');
+var docs_handler = express.static(__dirname + '/../swagger-ui/');
 app.get(/^\/docs(\/.*)?$/, function(req, res, next) {
   if (req.url === '/docs') { // express static barfs on root url w/o trailing slash
     res.writeHead(302, { 'Location' : req.url + '/' });
